Move ContractSourceManager to its own file
